モデルテスト合格

+++ db/migrate/20260104083336_remove_price_from_orders.rb
@@ -1,5 +1,8 @@
 class RemovePriceFromOrders < ActiveRecord::Migration[7.1]
   def change
-    remove_column :orders, :price, :integer
+    # column_exists? で確認してから削除するように変更
+    if column_exists?(:orders, :price)
+      remove_column :orders, :price, :integer
+    end
   end
 end
